+++ fsvolumecontents/changejournalview.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "pagewbdbase.h"
-#include "page_root.h"
-#include "page_changejournal.h"
-#include "interface.h"
-
-HRESULT CreateChangeJournalObject(HINSTANCE hInstance,IViewBaseWindow **pObject);