+++ fsvolumecontents/changejournalwindow.h
@@ -1,5 +1,0 @@
-// 2024.04.15
-#pragma once
-#include "basewindow.h"
-extern HWND CreateChangeJournalWindow(HWND hWnd);
-