+++ fsvolumecontents/dialogs.h
@@ -1,4 +1,0 @@
-#pragma once
-
-HRESULT GotoDirectoryDialog(HWND hWnd,PWSTR *ppszNewPath);
-