+++ fsvolumecontents/filesview.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "pagewbdbase.h"
-#include "page_root.h"
-#include "page_filelist.h"
-#include "interface.h"
-
-HRESULT CreateFileListBaseObject(HINSTANCE hInstance,IViewBaseWindow **pObject);