+++ fsvolumecontents/fileswindow.h
@@ -1,4 +1,0 @@
-// 2023.06.29
-#pragma once
-#include "basewindow.h"
-extern HWND CreateDirectoryFilesWindow(HWND hWnd);