+++ fsvolumecontents/fsvolumecontents.h
@@ -1,9 +1,0 @@
-#pragma once
-
-EXTERN_C
-HWND
-WINAPI
-CreateVolumeContentsBrowserWindow(
-	HWND hwnd,
-	UINT ConsoleId
-	);