+++ fsvolumewalker/volumewalker.cpp
@@ -29,6 +29,10 @@
 #define _ENABLE_SINGLETON_CONTENTS_BROWSER_WINDOW  1
 #endif
 
+#ifndef _ENABLE_VOLUME_CONTENTS
+#define _ENABLE_VOLUME_CONTENTS  0
+#endif
+
 static HINSTANCE g_hInst = NULL;
 static HWND g_hWndMain = NULL;
 static HWND g_hWndMDIClient = NULL;
@@ -302,7 +306,7 @@ static MDICHILDFRAMETABLE table[]= {
 	{VOLUME_CONSOLE_MOUNTEDDEVICE,     0},
 	{VOLUME_CONSOLE_DOSDRIVELIST,      0},
 	{VOLUME_CONSOLE_FILTERDRIVER,      0},
-#if _ENABLE_SINGLETON_CONTENTS_BROWSER_WINDOW
+#if _ENABLE_VOLUME_CONTENTS && _ENABLE_SINGLETON_CONTENTS_BROWSER_WINDOW
 	{VOLUME_CONSOLE_FILES,             0}, // todo: currently, singleton console.
 	{VOUUME_CONSOLE_CHANGE_JOURNAL,    0}, // todo: currently, singleton console.
 #endif
@@ -393,6 +397,7 @@ VOID MakeConsoleGUID(LPGUID pwndGuid,UINT ConsoleId,OPEN_MDI_CHILDFRAME_PARAM *O
 		case VOLUME_CONSOLE_FILTERDRIVER:
 			Guid.Data1 = (ULONG)ConsoleId;
 			break;
+#if _ENABLE_VOLUME_CONTENTS
 #if _ENABLE_SINGLETON_CONTENTS_BROWSER_WINDOW
 		case VOLUME_CONSOLE_FILES:
 		case VOUUME_CONSOLE_CHANGE_JOURNAL:
@@ -406,6 +411,7 @@ VOID MakeConsoleGUID(LPGUID pwndGuid,UINT ConsoleId,OPEN_MDI_CHILDFRAME_PARAM *O
 			Guid.Data2 = 0x1;
 			GetSystemTimeAsFileTime( (LPFILETIME)Guid.Data4 );
 			break;
+#endif
 #endif
 		case VOLUME_CONSOLE_VOLUMEINFORMAION:
 		case VOLUME_CONSOLE_PHYSICALDRIVEINFORMAION:
@@ -428,8 +434,10 @@ BOOL IsVolumeNameRequiredConsole(LPGUID pwndGuid,UINT ConsoleId)
 
 	switch( ConsoleId )
 	{
+#if _ENABLE_VOLUME_CONTENTS
 		case VOLUME_CONSOLE_FILES:
 		case VOUUME_CONSOLE_CHANGE_JOURNAL:
+#endif
 		case VOLUME_CONSOLE_VOLUMEINFORMAION:
 		case VOLUME_CONSOLE_PHYSICALDRIVEINFORMAION:
 		case VOLUME_CONSOLE_DISKLAYOUT:
@@ -453,8 +461,10 @@ PCWSTR GetConsoleTitle(UINT ConsoleTypeId)
 		{VOLUME_CONSOLE_MOUNTEDDEVICE,      L"Mounted Devices"},
 		{VOLUME_CONSOLE_DOSDRIVELIST,       L"Dos Drives"},
 		{VOLUME_CONSOLE_FILTERDRIVER,       L"Minifilter Driver"},
+#if _ENABLE_VOLUME_CONTENTS
 		{VOLUME_CONSOLE_FILES,              L"Volume Contents Browser"},
 		{VOUUME_CONSOLE_CHANGE_JOURNAL,     L"Volume Change Journal Browser"},
+#endif
 	};
 
 	PCWSTR pszTitle = L"";
@@ -488,10 +498,12 @@ HICON GetConsoleIcon(UINT ConsoleTypeId,PCWSTR pszInitialPath)
 		hIcon = (HICON)LoadImage(hmod,MAKEINTRESOURCE(67),IMAGE_ICON,16,16,LR_DEFAULTCOLOR);
 		FreeLibrary(hmod);
 	}
+#if _ENABLE_VOLUME_CONTENTS
 	else if( ConsoleTypeId == VOLUME_CONSOLE_FILES )
 	{
 		hIcon = GetShellStockIcon(SIID_DRIVEFIXED); // set temporary icon
 	}
+#endif
 	else
 	{
 		hIcon = GetShellStockIcon(SIID_DRIVEFIXED);
@@ -712,8 +724,11 @@ HWND OpenMDIChild(HWND hWnd,UINT ConsoleTypeId,LPGUID pwndGuid,OPEN_MDI_CHILDFRA
 		ConsoleTypeId == VOLUME_CONSOLE_SIMPLEHEXDUMP )
 #else
 		ConsoleTypeId == VOLUME_CONSOLE_SIMPLEHEXDUMP ||
+#if _ENABLE_VOLUME_CONTENTS
 		ConsoleTypeId == VOLUME_CONSOLE_FILES ||
-		ConsoleTypeId == VOUUME_CONSOLE_CHANGE_JOURNAL )
+		ConsoleTypeId == VOUUME_CONSOLE_CHANGE_JOURNAL
+#endif
+		)
 #endif
 
 	{
@@ -741,18 +756,20 @@ HWND OpenMDIChild(HWND hWnd,UINT ConsoleTypeId,LPGUID pwndGuid,OPEN_MDI_CHILDFRA
 
 		SendMessage(g_hWndMDIClient,WM_MDIACTIVATE,(WPARAM)hwndChildFrame,0);
 
-		if( VOUUME_CONSOLE_CHANGE_JOURNAL == ConsoleTypeId )
-		{
-			SendContentsBrowserChangeJournalVolumeOrFileName(hwndChildFrame,pd->hWndView,pOpenParam,FALSE);
-		}
-		if( VOLUME_CONSOLE_FILES == ConsoleTypeId )
-		{
-			SendContentsBrowserFileListPath(hwndChildFrame,pd->hWndView,pOpenParam,FALSE);
-		}
-		else if( VOLUME_CONSOLE_SIMPLEHEXDUMP == ConsoleTypeId && pOpenParam )
+		if( VOLUME_CONSOLE_SIMPLEHEXDUMP == ConsoleTypeId && pOpenParam )
 		{
 			SendHexDumpInformation(hwndChildFrame,pd->hWndView,pOpenParam);
 		}
+#if _ENABLE_VOLUME_CONTENTS
+		else if( VOUUME_CONSOLE_CHANGE_JOURNAL == ConsoleTypeId )
+		{
+			SendContentsBrowserChangeJournalVolumeOrFileName(hwndChildFrame,pd->hWndView,pOpenParam,FALSE);
+		}
+		else if( VOLUME_CONSOLE_FILES == ConsoleTypeId )
+		{
+			SendContentsBrowserFileListPath(hwndChildFrame,pd->hWndView,pOpenParam,FALSE);
+		}
+#endif
 
 		if( bMaximized )
 		{
@@ -820,6 +837,7 @@ HWND OpenMDIChild(HWND hWnd,UINT ConsoleTypeId,LPGUID pwndGuid,OPEN_MDI_CHILDFRA
 
 				SendHexDumpInformation(hwndMDIChild,pd->hWndView,pOpenParam);
 			}
+#if _ENABLE_VOLUME_CONTENTS
 			else if( VOUUME_CONSOLE_CHANGE_JOURNAL == ConsoleTypeId )
 			{
 				pd->hWndView = CreateVolumeContentsBrowserWindow(hwndMDIChild,ConsoleTypeId);
@@ -840,6 +858,7 @@ HWND OpenMDIChild(HWND hWnd,UINT ConsoleTypeId,LPGUID pwndGuid,OPEN_MDI_CHILDFRA
 
 				SendContentsBrowserFileListPath(hwndMDIChild,pd->hWndView,pOpenParam,TRUE);
 			}
+#endif
 			else
 			{
 				pd->hWndView = CreateVolumeConsoleWindow(hwndMDIChild,ConsoleTypeId,&param);
@@ -1487,7 +1506,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 				pcv = GetConsoleId(hwndViewWindow);
 				if( pcv == NULL )
 					return 0;
-
+#if _ENABLE_VOLUME_CONTENTS
 				switch( pcv->wndId )
 				{
 					case VOLUME_CONSOLE_FILES:
@@ -1522,6 +1541,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 						return 0;
 					}
 				}
+#endif // _ENABLE_VOLUME_CONTENTS
 			}
 			return 0;
 		}